Shuffling presets and plugins definition location
Moved presets and plugins from web pack configurations to .babelrc
Mainly this step has done so that the required transformations are
available for the unit testing step too. Initially with the presence in
the web pack configuration there where failures in unit tests where the
class transformation and other plugin related stuff where not working as
expected.